+++ cmd/queue-control-plane/workloads_test.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"testing"
 
+	controlkubernetes "github.com/faustbrian/go-queue-control-plane/adapters/kubernetes"
 	"github.com/faustbrian/go-queue-control-plane/control"
-	controlkubernetes "github.com/faustbrian/go-queue-control-plane/kubernetes"
 	kubernetesclient "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/fake"
 	"k8s.io/client-go/rest"