+++ dataplane/goqueue_integration_test.go
@@ -10,11 +10,11 @@ import (
 	"github.com/alicebob/miniredis/v2"
 	goqueue "github.com/faustbrian/go-queue"
 	controlplane "github.com/faustbrian/go-queue-control-plane"
+	redisstream "github.com/faustbrian/go-queue/adapters/redisstream"
 	"github.com/faustbrian/go-queue/core"
 	"github.com/faustbrian/go-queue/job"
 	queue "github.com/faustbrian/go-queue/management"
 	"github.com/faustbrian/go-queue/managementhttp"
-	redisstream "github.com/faustbrian/go-queue/redisstream"
 )
 
 func TestRedisDeadLettersFlowThroughControlPlaneContracts(t *testing.T) {